100
solution #1 : read, map every word and store it's freq. iterate every word and see if one of it's shifts is in the map.
solution #2 : read, map the normlaization of every word and stroe it's freq and the norm. form in the reading vector. iterate the vector and see if the the current normalization is more than 2 times in the mapping.